First pass at go wrapper
