+++ cmd/launcher/svc_windows.go
@@ -15,6 +15,7 @@ import (
 	"github.com/go-kit/kit/log/level"
 	"github.com/kolide/kit/logutil"
 	"github.com/kolide/kit/version"
+	"github.com/kolide/launcher/ee/gowrapper"
 	"github.com/kolide/launcher/pkg/autoupdate"
 	"github.com/kolide/launcher/pkg/contexts/ctxlog"
 	"github.com/kolide/launcher/pkg/launcher"
@@ -183,23 +184,7 @@ func (w *winSvc) Execute(args []string, r <-chan svc.ChangeRequest, changes chan
 
 	ctx = ctxlog.NewContext(ctx, w.logger)
 
-	go func() {
-		// Log panics from runLauncher
-		defer func() {
-			if r := recover(); r != nil {
-				w.systemSlogger.Log(ctx, slog.LevelInfo,
-					"panic occurred in runLauncher",
-					"err", r,
-				)
-				if err, ok := r.(error); ok {
-					w.systemSlogger.Log(ctx, slog.LevelInfo,
-						"panic stack trace",
-						"stack_trace", fmt.Sprintf("%+v", errors.WithStack(err)),
-					)
-				}
-			}
-		}()
-
+	gowrapper.Go(w.systemSlogger.Logger, func() {
 		err := runLauncher(ctx, cancel, w.slogger, w.systemSlogger, w.opts)
 		if err != nil {
 			w.systemSlogger.Log(ctx, slog.LevelInfo,
@@ -221,7 +206,13 @@ func (w *winSvc) Execute(args []string, r <-chan svc.ChangeRequest, changes chan
 		)
 		changes <- svc.Status{State: svc.Stopped, Accepts: cmdsAccepted}
 		os.Exit(0)
-	}()
+	}, func(r any) {
+		w.systemSlogger.Log(ctx, slog.LevelError,
+			"exiting after runLauncher panic",
+			"err", r,
+		)
+		os.Exit(1)
+	})
 
 	for {
 		select {
